Fix the bugs:
1.
Before: The music service won't start in certain cases.
After: The music service will start once the song is playing.

2.
Before: When we first launch the app, the pager can swipe to whatever page we want. However, when we press the play button after swiping to a certain page, it will play the first song instead of the intended song.
After: When we first launch the app, we can't swipe to the desired song. Instead, the song will play as soon as you swipe the pager player."